Description
A simple fighting game.
Both player have a certain level of health and a chance to strike. The chance is determined by a random number, and if it is greater than opponents, player does a hit, and the strength of it depends on how close the chance was to the maximum possible value. With each move, health decreases, and the game ends when the health of one of the parties is less than or equal to zero.